Normal map

+++ src/Math.hpp
@@ -3,8 +3,8 @@
 #define PI 3.1415926
 
 glm::vec3 get_refract(
-              const glm::vec3 & normal, 
-              const glm::vec3 & direction,
+              const glm::dvec3 & normal, 
+              const glm::dvec3 & direction,
               double n1, 
               double n2
 );
